shop function module
可読性を高めるため、API側と名称を統一する。
